rewrite for v1.0.0
Make several API changes which were pending for a long time.
Most of the older APIs have been deprecated to allow for better things in the future.
One important advantage is that now Row headers are available.
Foundation for merging cells has been added which is currently planned to be used but not limited to spanning cells support.

* Fixes #57 (ANSI regex issue with Blessings)
* Fixes #65 (separate header alignment)
* Progress on #74 . (cell spanning) (targeted for v1.1.0)
* Fixes #89 , #27 , #15 (A better interface for auto vs manual width)
* Fixes #90 (Row headers)